fix: revert 051 lots column edit and fix 083 comment semicolon parse error
Migration 051 had lots UInt8→UInt32 edited in-place after running (handled
by migration 084). Reverted to the originally-run version to match DB checksum.

Migration 083 had two bugs that caused migrate.py to fail:
- Inline comment example ended with `;\n` — the splitter treats it as a
  statement boundary, sending ClickHouse a comment-only empty query.
- Adjacent string literal in INSERT VALUES (Python syntax, not ClickHouse SQL).
Both fixed. Tables already existed from a partial prior attempt; migrations now
run cleanly to completion (083–087 all green).

+++ clickhouse/migrations/051_create_trade_recommendations.sql
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS analysis.strategy_simulation
     short_n          UInt8,
     wing_m           UInt8,
     confidence       Float32,
-    lots             UInt32,
+    lots             UInt8,
     lot_size         UInt16,
     pnl_pts          Float32,
     pnl_amount       Float32,
